Fix build
The error was caused by a classloading loop.
