+++ fabric-renderer-indigo/src/client/java/net/fabricmc/fabric/impl/client/indigo/renderer/mesh/EncodingFormat.java
@@ -26,7 +26,6 @@ import net.minecraft.util.math.MathHelper;
 
 import net.fabricmc.fabric.api.renderer.v1.mesh.QuadView;
 import net.fabricmc.fabric.api.renderer.v1.model.ModelHelper;
-import net.fabricmc.fabric.impl.client.indigo.renderer.IndigoRenderer;
 import net.fabricmc.fabric.impl.client.indigo.renderer.helper.GeometryHelper;
 import net.fabricmc.fabric.impl.client.indigo.renderer.material.MaterialViewImpl;
 import net.fabricmc.fabric.impl.client.indigo.renderer.material.RenderMaterialImpl;
@@ -102,27 +101,6 @@ public final class EncodingFormat {
 
 	static {
 		Preconditions.checkArgument(TOTAL_BIT_LENGTH <= 32, "Indigo header encoding bit count (%s) exceeds integer bit length)", TOTAL_STRIDE);
-	}
-
-	static final int[] EMPTY = new int[TOTAL_STRIDE];
-	/** Used for quick clearing of quad buffers. Implicitly has invalid geometry. */
-	static final int[] DEFAULT = EMPTY.clone();
-
-	static {
-		MutableQuadViewImpl quad = new MutableQuadViewImpl() {
-			@Override
-			protected void emitDirectly() {
-				// This quad won't be emitted. It's only used to configure the default quad data.
-			}
-		};
-
-		// Start with all zeroes
-		quad.data = DEFAULT;
-		// Apply non-zero defaults
-		quad.color(-1, -1, -1, -1);
-		quad.cullFace(null);
-		quad.material(IndigoRenderer.STANDARD_MATERIAL);
-		quad.tintIndex(-1);
 	}
 
 	public static int bitMask(int bitLength, int bitOffset) {
